+++ CHANGELOG.md
@@ -6,6 +6,7 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 ### Added
 - Diffing of Reviewer Candidates
 - Authors will see the feedback form whenever they submit a paper
+- Add manifest to Apex export
 ### Changed
 - Update RTC and FTC text.
 - Billing task institutions and countries synced with EM